Добавит возможность создавать новые точки маршрута
