+++ public/index.html
@@ -32,7 +32,8 @@
             </div>
           </div>
 
-          <button class="trip-main__event-add-btn  btn  btn--big  btn--yellow" type="button">New event</button>
+          <button class="trip-main__event-add-btn  btn  btn--big  btn--yellow" type="button">
+            New event</button>
         </div>
       </div>
     </header>
